requirements.txt
matplotlib
tensorflow
ipykernel
scikit-learn
Flask
Pillow